requirements.txt
Django==1.11.2
django-allauth==0.32.0
django-crispy-forms==1.6.1
django-custom-user==0.7
django-guardian==1.4.8
django-invitations==1.9.1
django-webpack-loader==0.5.0
uwsgi==2.0.15
psycopg2==2.7.1
ipython==6.1.0
numpy==1.13.0
djangorestframework==3.6.3
python-magic==0.4.12
raven==6.1.0
xlrd==1.0.0
XlsxWriter==0.9.6
pandas==0.20.2
plotly==2.0.10
scipy==0.19.0
seaborn==0.7.1
tables==3.4.2
